Begin the 2.0 switch to nightmare
