+++ package.json
@@ -4,7 +4,7 @@
   "description": "Curriculum crawler checks for broken links in the App Academy curriculum.",
   "main": "index.js",
   "scripts": {
-    "start": "mkdir -p dist && babel index.js --out-file dist/index.js && node ./dist/index.js",
+    "start": "mkdir -p dist && babel index.js --out-file dist/index.js && node --harmony ./dist/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
@@ -13,13 +13,9 @@
     "babel-cli": "^6.22.2",
     "babel-loader": "^6.2.10",
     "babel-preset-es2015": "^6.22.0",
-    "casperjs": "^1.1.3",
     "eslint-config-airbnb": "^14.0.0",
-    "phantom": "2",
-    "phantomjs": "^2.1.7",
+    "nightmare": "^2.9.1",
     "readline-sync": "^1.4.6",
-    "spooky": "^0.2.5",
-    "tiny-jsonrpc": "^2.0.1",
     "webpack": "^2.2.1"
   },
   "devDependencies": {
